Move `dead_end/internals` to `dead_end`
This file was created so that there could be an "fyi" interface that told people  to run the CLI command, as well as a "auto" interface that fired the search functionality.

I did this for safety reasons originally. Since then, the reliability has increased. We have deprecated and then removed the `dead_end/fyi` interface and can now also get rid of this internals file. That allows us to have the core functionality where you would expect it.
